Handling missing data
Some SMAP scenes are missing (e.g., 2015-05-13). Previously, searching for smap data using find_smap would raise an error when no data existed. As suggested in #24, this is annoying when trying to search for data over a long time period.

To better handle missing data, now find_smap() will raise a warning instead of an error, and insert a row with an NA value for the file and location of the file, so that users have a record of which date(s) lack data.

This pull request also adds functionality to download_smap to handle the case when users try to download files that are missing. If a user passes a data frame with NA values, download_smap will raise an error and explain that it cannot accept NA values (i.e., it cannot download data that does not exist).

Last, this PR also updates data versions for the examples.