Adding verbose argument to download_smap()
This adds an argument to download_smap() that prints which files
are being downloaded, to reassure users that things are really
happening.

Also, this includes an unrelated improvement to the documentation
for find_smap() that points users to the "Details" section for a
list of data set ids that are supported by smapr.

Solves #44